requirements.txt
playwright
python-docx
pdfplumber
google-generativeai
groq
pyyaml
python-dotenv
